Rename modules to match Belgium
