+++ python/popgetter/assets/scotland/census_tables.py
